requirements/lint.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --no-annotate requirements/lint.in
#
black==22.6.0
click==8.1.3
flake8==4.0.1
flake8-force==0.0.1
flake8-plugin-utils==1.3.2
flake8-pytest-style==1.6.0
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pyupgrade==2.37.3
tokenize-rt==4.2.1
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
zipp==3.8.1
